requirements.txt
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
pyyaml>=6.0
tqdm>=4.65.0
bert-score>=0.3.13
openai>=1.3.0
anthropic>=0.18.0
google-generativeai>=0.3.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jsonlines>=3.1.0
simhash>=2.1.1
nltk>=3.8.1
rouge-score>=0.1.2
rank_bm25>=0.2.2
datasets>=2.14.0
pytest>=7.4.0
